Reorganize code and add Client metrics
