+++ unifi_console/metadata.csv
@@ -27,3 +27,14 @@ unifi.device.uplink.tx_dropped,gauge,,,,uplink number of TX dropped,1,unifi_cons
 unifi.device.uplink.tx_errors,gauge,,,,uplink number of TX errors,1,unifi_console,uplink tx errors,
 unifi.device.uplink.tx_packets,gauge,,packet,,uplink transmitted packets,1,unifi_console,uplink tx packets,
 unifi.device.uplink.up,gauge,,,,uplink status,0,unifi_console,status,
+unifi.client.noise,gauge,,,,client noise in dBm,0,unifi_console,noise,
+unifi.client.rx_bytes,gauge,,byte,,client received bytes,1,unifi_console,client rx bytes,
+unifi.client.rx_packets,gauge,,packet,,client received packets,1,unifi_console,client rx packets,
+unifi.client.rx_rate,gauge,,byte,second,client received rate,1,unifi_console,client rx rate,
+unifi.client.satisfaction,gauge,,percent,,client satisfaction,1,unifi_console,client satisfaction,
+unifi.client.signal,gauge,,,,client signal in dBm,1,unifi_console,client signal,
+unifi.client.tx_bytes,gauge,,byte,,client transmitted bytes,1,unifi_console,client tx bytes,
+unifi.client.tx_packets,gauge,,packet,,client transmitted packets,1,unifi_console,client tx packets,
+unifi.client.tx_rate,gauge,,byte,second,client transmitted rate,1,unifi_console,client tx rate,
+unifi.client.tx_retries,gauge,,,,client transmitted retry,1,unifi_console,client transmitted retry,
+unifi.client.uptime,gauge,,second,,client uptime,1,unifi_console,client uptime,
