+++ unifi_console/tests/fixtures/client_metrics.json
@@ -1,0 +1,81 @@
+{
+    "meta": {
+        "rc": "ok"
+    },
+    "data": [
+        {
+            "site_id": "61964cf2f16a5b3422373c85",
+            "assoc_time": 1655898552,
+            "latest_assoc_time": 1655898622,
+            "oui": "",
+            "user_id": "6196923af16a5b038fce91bb",
+            "_id": "6196923af16a5b038fce91bb",
+            "dev_id_override": 4426,
+            "disconnect_timestamp": 1655898538,
+            "fingerprint_override": true,
+            "first_seen": 1637257786,
+            "is_guest": false,
+            "is_wired": false,
+            "last_seen": 1655916558,
+            "mac": "11:11:11:11:11:11",
+            "name": "My device",
+            "noted": true,
+            "use_fixedip": false,
+            "usergroup_id": "",
+            "hostname": "My-device",
+            "wlanconf_id": "61965b18f16a5b03c811329c",
+            "_uptime_by_uap": 17936,
+            "_last_seen_by_uap": 1655916558,
+            "_is_guest_by_uap": false,
+            "ap_mac": "11:11:11:11:11:11",
+            "channel": 44,
+            "radio": "na",
+            "radio_name": "rai0",
+            "essid": "My Network",
+            "bssid": "11:11:11:11:11:11",
+            "powersave_enabled": true,
+            "is_11r": false,
+            "user_group_id_computed": "61964cfff16a5b3422373c98",
+            "anomalies": 0,
+            "anon_client_id": "b87303274660f4ba1c7967c064b172",
+            "ccq": 0,
+            "dhcpend_time": 0,
+            "idletime": 1,
+            "ip": "1.1.1.1",
+            "noise": -96,
+            "nss": 2,
+            "rx_rate": 6000,
+            "rssi": 27,
+            "satisfaction": 97,
+            "satisfaction_now": 96,
+            "satisfaction_real": 97,
+            "satisfaction_reason": 16,
+            "signal": -69,
+            "tx_mcs": 5,
+            "tx_power": 0,
+            "tx_rate": 240000,
+            "hostname_source": "uap",
+            "vlan": 0,
+            "radio_proto": "ac",
+            "satisfaction_avg": {
+                "total": 103357,
+                "count": 1051
+            },
+            "uptime": 18006,
+            "tx_bytes": 1817775057,
+            "rx_bytes": 18740395,
+            "tx_packets": 1477832,
+            "rx_packets": 108960,
+            "bytes-r": 2,
+            "tx_bytes-r": 0,
+            "rx_bytes-r": 2,
+            "tx_retries": 101645,
+            "wifi_tx_attempts": 1579439,
+            "authorized": true,
+            "qos_policy_applied": true,
+            "roam_count": 2,
+            "fingerprint_source": 0
+        }
+        
+    ]
+}
