+++ unifi_console/tests/fixtures/client_metrics_values.json
@@ -1,0 +1,167 @@
+[
+    {
+        "name": "unifi.client.noise",
+        "value": -96,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.rx_bytes",
+        "value": 18740395,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.rx_packets",
+        "value": 108960,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.rx_rate",
+        "value": 6000,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.satisfaction",
+        "value": 97,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.signal",
+        "value": -69,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.tx_bytes",
+        "value": 1817775057,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.tx_packets",
+        "value": 1477832,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.tx_rate",
+        "value": 240000,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.tx_retries",
+        "value": 101645,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    },
+    {
+        "name": "unifi.client.uptime",
+        "value": 18006,
+        "tags": [
+            "id:6196923af16a5b038fce91bb",
+            "radio_name:rai0",
+            "device:My device",
+            "channel:44",
+            "radio:na",
+            "radio_proto:ac",
+            "oui:",
+            "url:https://FAKE",
+            "site:default"
+        ]
+    }
+]
