+++ unifi_console/tests/mocked_api.py
@@ -3,7 +3,9 @@ import os
 
 from tests.common import HERE
 
-from datadog_checks.unifi_console.types import ControllerInfo, DeviceInfo
+from datadog_checks.unifi_console.client_info import ClientInfo
+from datadog_checks.unifi_console.device_info import DeviceInfo
+from datadog_checks.unifi_console.types import ControllerInfo
 
 
 class MockedAPI(object):
@@ -26,3 +28,13 @@ class MockedAPI(object):
                 devices.append(DeviceInfo(obj))
 
             return devices
+
+    def get_clients_info(self):
+        with open(os.path.join(HERE, "fixtures", "client_metrics.json")) as f:
+
+            resp = json.load(f)
+            clients = []
+            for obj in resp["data"]:
+                clients.append(ClientInfo(obj))
+
+            return clients
